+++ src/plugins/dam/components/image-asset-picker-field/index.jsx
@@ -1,17 +1,22 @@
-import PropTypes from 'prop-types';
-import React from 'react';
+import { Button, FormGroup, FormText, Label, Icon, RouterLink } from '@triniti/admin-ui-plugin/components';
+import { connect } from 'react-redux';
 import damUrl from '@triniti/cms/plugins/dam/utils/damUrl';
+import ImageAssetPickerModal from '@triniti/cms/plugins/dam/components/image-asset-picker-modal';
 import ImageAssetV1Mixin from '@triniti/schemas/triniti/dam/mixin/image-asset/ImageAssetV1Mixin';
 import Message from '@gdbots/pbj/Message';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
-import { Button, FormGroup, FormText, Label } from '@triniti/admin-ui-plugin/components';
-import ImageAssetPickerModal from '@triniti/cms/plugins/dam/components/image-asset-picker-modal';
+import pbjUrl from '@gdbots/pbjx/pbjUrl';
+import PropTypes from 'prop-types';
+import React from 'react';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
+import selector from './selector';
 
 const imageType = ImageAssetV1Mixin.findOne().getCurie().getMessage();
 
 class ImageAssetPickerField extends React.Component {
   static propTypes = {
     areLinkedImagesAllowed: PropTypes.bool,
+    getNode: PropTypes.func.isRequired,
     input: PropTypes.shape({
       onChange: PropTypes.func,
       value: PropTypes.oneOfType([
@@ -78,12 +83,41 @@ class ImageAssetPickerField extends React.Component {
       label,
       meta: { error },
       node,
+      getNode,
     } = this.props;
+
+    const image = getNode(value);
 
     return (
       <div className="mb-4">
         <FormGroup className="mb-0">
           <Label>{label}</Label>
+          {image && (
+          <>
+            <RouterLink to={pbjUrl(image, 'cms')}>
+              <Button id={`view-${image.get('_id')}`} size="xs" color="hover" radius="circle" className="mb-1 ml-2">
+                <Icon imgSrc="eye" alt="view" />
+              </Button>
+              <UncontrolledTooltip placement="auto" target={`view-${image.get('_id')}`}>View</UncontrolledTooltip>
+            </RouterLink>
+            <RouterLink to={`${pbjUrl(image, 'cms')}/edit`}>
+              <Button id={`edit-${image.get('_id')}`} size="xs" color="hover" radius="circle" className="mb-1 mx-1">
+                <Icon imgSrc="pencil" alt="edit" />
+              </Button>
+              <UncontrolledTooltip placement="auto" target={`edit-${image.get('_id')}`}>Edit</UncontrolledTooltip>
+            </RouterLink>
+            <a
+              href={damUrl(NodeRef.fromNode(image))}
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <Button id={`open-in-new-tab-${image.get('_id')}`} size="xs" color="hover" radius="circle" className="mb-1">
+                <Icon imgSrc="external" alt="open" />
+              </Button>
+              <UncontrolledTooltip placement="auto" target={`open-in-new-tab-${image.get('_id')}`}>Open in new tab</UncontrolledTooltip>
+            </a>
+          </>
+          )}
           {
             value && (
               <img
@@ -127,4 +161,4 @@ class ImageAssetPickerField extends React.Component {
     );
   }
 }
-export default ImageAssetPickerField;
+export default connect(selector)(ImageAssetPickerField);