+++ src/plugins/ncr/screens/node/index.jsx
@@ -4,6 +4,7 @@ import Collaborators from '@triniti/cms/plugins/raven/components/collaborators';
 import get from 'lodash/get';
 import getUserConfirmation from '@triniti/admin-ui-plugin/utils/getUserConfirmation';
 import isEmpty from 'lodash/isEmpty';
+import LabelsForm from '@triniti/cms/plugins/ncr/components/labels-form';
 import NodeLock from '@triniti/cms/plugins/ncr/components/node-lock';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 import NodeStatus from '@triniti/cms/plugins/ncr/components/node-status';
@@ -320,6 +321,24 @@ export default class AbstractNodeScreen extends React.Component {
     return <Chat key="chat" nodeRef={nodeRef} />;
   }
 
+  renderLabelsForm() {
+    const { delegate, isSaveDisabled } = this.props;
+    const node = delegate.getNode();
+
+    if (!node || !node.schema().hasMixin('gdbots:common:mixin:labelable')) {
+      return null;
+    }
+
+    return (
+      <LabelsForm
+        disabled={!isSaveDisabled}
+        key="labels"
+        node={delegate.getNode()}
+        schemas={delegate.getSchemas()}
+      />
+    );
+  }
+
   renderNodeLock() {
     const {
       delegate,
@@ -343,16 +362,16 @@ export default class AbstractNodeScreen extends React.Component {
     );
   }
 
-  renderPreviewButtons() {
-    const { delegate } = this.props;
-    const node = delegate.getNode();
-    return <PreviewButtons key="PreviewButtons" node={node} />;
-  }
-
   renderNodeStatus() {
     const { delegate } = this.props;
     const node = delegate.getNode();
     return node ? <NodeStatus key="NodeStatus" node={node} /> : null;
+  }
+
+  renderPreviewButtons() {
+    const { delegate } = this.props;
+    const node = delegate.getNode();
+    return <PreviewButtons key="PreviewButtons" node={node} />;
   }
 
   renderPublishForm() {
@@ -383,6 +402,7 @@ export default class AbstractNodeScreen extends React.Component {
       this.renderPreviewButtons(),
       this.renderNodeStatus(),
       this.renderPublishForm(),
+      this.renderLabelsForm(),
       this.renderChat(),
     ];
   }