+++ src/plugins/pbjx/reducers/index.test.js
@@ -21,6 +21,7 @@ test('Pbjx:reducer:response:clearResponse', (t) => {
   let expected = fakeState;
   t.deepEqual(actual, expected, 'it should return correct state if no provided curie');
 
+
   fakeState = {
     test: {
       'whatevs-slot-here': {
@@ -59,7 +60,6 @@ test('Pbjx:reducer:response:clearResponse', (t) => {
   expected = fakeState;
   t.deepEqual(actual, expected, 'it should return correct state if no response found');
 
-
   fakeState = {
     test: {
       'test-slot': {
@@ -71,7 +71,6 @@ test('Pbjx:reducer:response:clearResponse', (t) => {
   actual = reducer(fakeState, clearResponse(curie));
   expected = fakeState;
   t.deepEqual(actual, expected, 'it should return correct state if no channel and slot found');
-
 
   fakeState = {
     test: {